requirements.txt
flask==2.3.2
flask-bcrypt==1.0.1
mysql-connector-python==8.0.33
werkzeug==2.3.6
pillow==9.5.0 